Kernel/Clock: Added `TSC` clock device based on cpu timestamp
This is very efficient clock, but we need to calibrate it as it is based on CPU clock, and not on seconds.

It also has an issue where, the clock frequency of the CPU might change,
and we need to somehow re_calibrate when needed. For now we only do it
once.