Kernel/Process: fix bug in executing some SSE instructions
Earlier we changed the userspace target in `rust` to enable `SSE`,
and it was working great, until we found a bug.

Some programs would crash with `#GP` exception from instructions such as
`MOVAPS`.

After looking around, looks like this instruction require at least
16-byte memory alignment. The value passed was the `rsp`, and it was
8-byte alignment for some reason.

After looking around more, looks like this is originating from the SYSV
ABI. SYSV ABI require the stack to be 16-byte aligned before the `call`
instruction, i.e. after the call instruction, it will be `16 byte + 8`
aligned.

And of course, the only place we control the stack is on process
startup.

I was able to also test that this is the cause, since adding another
argv value fixed it :D.

Similar issues I found that helped me find the answer:
https://github.com/phil-opp/blog_os/issues/179