Kernel/Memory: changed mentality of dealing with memory int types
Now, `virtual` addresses are `usize`, and `physica` addresses are `u64`, the size of a memory region is also `usize`, as it should be contained in virtual memory.

This is because:
- `usize` can be used later to implement `32bit` platforms easily.
- on x86 machines if `PG` is set, it can translate `32-bit` addresses into `40-bit` physical space.
So we can't use `usize` for that.

There are still some places uses `u64`, like the `ELF`, I'm reluctent to use `usize` there, as its part of the format. We will have to see how to support `32bit` applications loading later on.