Kernel/Filesystem: Don't read whole directory every time
Changed the `read_dir` so that we can read elements incrementally and
thus we can stop whenever we need.

I tried returning something that is `Iterator`, but its not possible due
to `object safety`, so a callback is the best we can use, but it works
well.

The still downside that I want to fix is the ability to continue reading
from where we left off, but for that, we need to save a state somehow,
and I think its not needed right now.