Kernel/Filesystem: Add filesystem mappings when reading directories
This mean that if we have two mappings: `/` and `/devices`, when reading
the directory `/` we should get `/devices`, previously we don't since
`/` will only query its children.

This only affects `Directory`, i.e. the user facing object, this is
important since the kernel always associate a `FileSystem` object
together with its `Node`, and here we violate that since this new `Node`
doesn't belong to the same `FileSystem` being queried.