Kernel/FAT: Fixed bugs when writing to disk
We need to keep which `cluster_index` in the current file, we are using,
instead of `previous_position`.

This is because, we can sometime write, and stop exactly at the boundary
between clusters, with the previous method, we won't know if the current
cluster includes the current position or not.

Also changed committing to disk so that we commit everything that has
changed so far, i.e. we won't commit the new file size until we are
writing the new file content.