Kernel/Filesystem+FAT: Added a way to create a node in a directory
This node is either a file or directory.

We create the mappings and dir entries, both long and normal and add
them to the directory.

We start with an entry of size=0 and cluster=0.

There is a bug in this is that creating a directory might lead us to the
root directory, which is probably we don't want.