Kernel/FAT: Use bitmap for dirty FAT sectors
FAT can be quite a large number of sectors, for my test for example
using qemu vvfat, its 500~ sectors, writing all of these in any change
is quite costly and not needed.

Used a bitmap, and its quite efficient, for example, for the above
number of sectors, we only need 500~ bits, which is not large.