xtask: Added `test` command for running kernel tests
It will compile a special elf and embed it into an iso and run it.

This is actually the main reason we switched to `xtask`, getting this
`elf` file from `cargo-make` was not possible, at least didn't find a
way to do that.

We could have used cargo runner field in .cargo/config.toml, but this is
for the `cargo test` command, and since we were using `cargo-make`, I
preferred to use one set of commands