Kernel/Scheduler: Use `BinaryHeap` as a priority queue
This method allow for quite fast access, the scheduler main loop will
just look at the top element, i.e. the element with highest
`priority_counter`. This counter is decremented then, the plan is to
have `PriorityLevel` for each process that will control how much this
value is decremented to control how often this process is run.

There are some info to be noted of:
- When reschedule_process, we should reset the counter to a reasonable
  value, because if imagine a process was sleeping for 100 seconds, it
  will continue running alone for the next 100 seconds which doesn't
  make much sense.

There are a lot of issues still in this, like:
- counter underflow (after long long time)
    - here, I'm planning to reset all processes, its a heavy process but
      not that much and is not done often