Kernel/Scheduler: Use `RefCell` for interior mutability
We want to be able to modify the Process easily, the issue with
`BinaryHeap` is that you can't get `iter_mut` since that would mess up
the ordering so much and they can't track it (compared to `peek_mut`).

So, here is a good place to use `RefCell`, where we can get inner
mutability safely.