Kernel/Fat: Use at least 1 cluster for file nodes
Previously, we would remove the first cluster from the file if the size
become or is `0`, which is more annoying to maintain and I think it
might create issues in the future.

Currently, when we create a file, the first cluster is allocated and
never removed until the file is removed.