Kernel/Tracing: Don't use `tracing` in `panic_handler`
`panic_handler` is very critical, and since tracing contain a lot of
functionalities (like accessing the filesystem, console, etc.), if one
of these panics, we will get into infinite loop and become harder to
debug.

So, this will stay as `println`, since it has less dependencies and has
fault tolerance with `ReMutex`.