Kernel/IDE: Fix bug in reading more than 255 sectors
For some reason, I think I just copied from the ATAPI code,
and in there, it accepts u16 value, but here, we were just
casting `u16` to `u8` which is very very bad of course.

It was resulting in part of the FAT table being missing and it was weird
behavior, when `next_cluster` function throws an error because the
filesystem is corrupted.

A good idea is to add some checks for the filesystem on load