Kernel/Clock: added `PIT` driver as alternative to `HPET`
If the machine doesn't have `HPET`, we will use the `PIT` instead.

Though, there are few issues with it now, for example, its has very small counter,
so it wraps around quickly, we need to fix the usage of the counter and make it better.

Also fixed a bug in `TSC` calibration where it will crash when using
`PIT` due to the small counter wrapping around. This is not a permanent
fix and we will need to change how this work next