Kernel/AML: improve displaying AML code using `fmt::Display`
I copied parts of how `DebugStruct` is implemented in the `std` library
in rust. Because we needed to have some kind of recursive structure.

Due to us not having access to internal methods of `Formatter`, we
needed to use another way.

There is also another part of this commit.
We fixed the parsing of `FieldDef` and `IndexFieldDef` and parsed the
`flags` field. It was a bit annoying to separate that so I included it
here in this PR.

There are a lot more to improve on this, but this will make printing the
data much much easier instead of worrying about `depth` and other items.