Kernel/AML: Moved `Buffer` into its own struct
We want to also support identifying `Buffer` macros, these are macros in
ASL that would generate Buffer, since we are dealing with AML directly,
we have them compiled to Buffers, and we don't know what the source code
actually contained. We can reverse the process, but we need to identify
the macro used based on context with other elements.