Kernel/Scheduler: make scheduler::shutdown flag global
This is needed because I tried implementing shutdown access from
userspace through the `/devices` filesystem, and issuing shutdown
command by writing to a specific file.

It works well except when calling `scheduler::stop_scheduler`, because
the `sys_write` syscall reads the files from the `Process` struct stored
inside the `SCHEDULER`, and when calling `stop_scheduler` we will still
be inside the `Mutex` of `SCHEDULER` and this causes double lock.

The fix is to have the `SHUTDOWN` flag be global atomic, easy to modify,
the scheduler will check if it needs to shutdown and does that and
continue the shutdown procedure as is.

Signed-off-by: Amjad Alsharafi <26300843+Amjad50@users.noreply.github.com>