Kernel/Misc: Added `macro_rules_attribute` to improve macro calling
The issue with using `testing::test!` is that formatting doesn't work as
it doesn't have much information on what's inside the macro.

This new macro makes it better in that regard.

I don't like using external crates if not essential, but we are planning
to use proc-macro to replace this anyway in the future.
(See https://github.com/Amjad50/Emerald/issues/117)

Signed-off-by: Amjad Alsharafi <26300843+Amjad50@users.noreply.github.com>