Kernel/Net: Start work on `E1000` driver, and now we can receive packets
We can see some broadcast ethernet packets, I'm not sure how to send
packets from my machine to this. For now it seems that's a bit tricky.

But at least we have something, we can try to send broadcast ethernet
later when implementing transmitting and then see what we get back.

Signed-off-by: Amjad Alsharafi <26300843+Amjad50@users.noreply.github.com>