Kernel/Net: Add `NetworkPacket::buildable()` for dynamic packet build
The issue with the previous method is that you need to know/predict what
headers there will be, which is unrealistic, so this is a better option.

tbh, this is probably not a very nice nor efficient method, but still
planning what to change it to.

For now its okay I guess

Signed-off-by: Amjad Alsharafi <26300843+Amjad50@users.noreply.github.com>